Address @mkoval's comments.
