+++ include/aikido/constraint/TestableOutcome.hpp
@@ -1,9 +1,7 @@
 #ifndef AIKIDO_CONSTRAINT_TESTABLEOUTCOME_HPP_
 #define AIKIDO_CONSTRAINT_TESTABLEOUTCOME_HPP_
 
-#include <sstream>
-#include <stdexcept>
-#include <typeinfo>
+#include <string>
 
 namespace aikido {
 namespace constraint {
@@ -28,24 +26,11 @@ public:
 /// pointers down to pointers for a derivative class. Mostly used in the
 /// isSatisfied methods of classes that inherit Testable.
 template <class Child>
-Child* dynamic_cast_if_present(TestableOutcome* outcome)
-{
-  if (!outcome)
-    return nullptr;
-
-  auto childPtr = dynamic_cast<Child*>(outcome);
-  if (!childPtr)
-  {
-    std::stringstream message;
-    message << "TestableOutcome pointer is not of type " << typeid(Child).name()
-            << ".";
-    throw std::invalid_argument(message.str());
-  }
-
-  return childPtr;
-}
+Child* dynamic_cast_if_present(TestableOutcome* outcome);
 
 } // namespace constraint
 } // namespace aikido
+
+#include "detail/TestableOutcome-impl.hpp"
 
 #endif // AIKIDO_CONSTRAINT_TESTABLEOUTCOME_HPP_
